Add type module
Move types into types.rs. Keep the parsing methods in parse.rs.

UIntType::to_simplicity(&self) -> UIntType::to_simplicity(self) to
satisfy clippy
